It would've been better to break this into multiple commits since there are a decent amount of changes.
In ClientSession, renamed _connectionClient to _client, and renamed 'TradingSessionController' to 'ClientSession' when it logs.

In MetaTraderClient, changed the _eventHandler field so it uses the IMTEventHandler interface instead of the MTEventHandler class. Also, modified the constructor so that the logger argument has a default value of null and when that is the case, it will use a ConsoleLogger. Then, changed some field assignments in the constructor, and usages, to use the _config properties instead of method arguments

In MetaTraderSession, commented out _instanceThread assignment because it is already assigned in the base class ClientSession. Also moved _client.Start() from the Run method overriden in the MetaTraderSession class to become the default implementation of the Run method in the base class ClientSession.

In MTConfiguration, added public properties SleepDelayMilliseconds, MaxRetryCommandsSeconds, LoadOrdersFromFile, & Verbose with corresponding private fields which have default values. This was done so that those variables could be moved out of the MetaTraderClient constructor where they were originally assigned.

Finally, I created a new assembly/project in the same solution called TradingAPIsUnitTests, and started to modify the existing unit test from Darwinex to worked with the renamed variables in my project. I will work more on unit testing now.