* In MetaTraderUnitTest:     - Re-styled the local fields to be explicitly private (although by default they were already private as no access modifiers were supplied), and prefixed with an underscore.     - Then, added a new private field called _testConfig, which is an MTConfiguration type that's initialized in the declaration by passing the 'mt4TestConfig.ini' filepath.     - Also renamed the client which MetaTraderUnitTest uses from dwx to testClient.
* In MTEventHandler, modified constructor so that the logger argument has a default value of null which means it will use a ConsoleLogger by default.

* In the TradingAPIsUnitTests folder, added mt4TestConfig.ini for initalizing the test MetaTraderClient.

Note: When I tried this using the mql server from the DWXConnect develop branch, none of the tests passed. When I tried with the mql server from the DWXConnect main branch, it passed 2/6 tests. I thought it might have been because of the market being closed over the weekend but now the market should still be open and it's still failing.
    * The tests it passed are:
        - TestGetHistoricData
        - TestSubscribeSymbolsBarData
    * The tests it failed are:
        - TestCloseAllOrders
        - TestCloseOrdersByMagic
        - TestCloseOrdersBySymbol
        - TestOpenModifyCloseOrder

I will have to spend more time debugging, and also create more test methods.